Update model.ex to allow binary parameter maps
I had an issue when trying to work with arc’s remote-files branch: I was getting a changeset message that my required ecto parameter could not be empty.

What I eventually determined to be the cause of my troubles was that the Dict.take was coming up empty as there was a discrepancy between string keys in the required/optional field module attributes, versus atom keys in the incoming params. 

I'm very new to Elixir so this code is probably less than ideal, but feel free to take a look and see if this makes sense.